Je suis leo et ceci est une usurpation d identite

+++ ping/src/main/java/fr/epita/assistants/ping/presentation/rest/FolderResource.java
@@ -3,40 +3,24 @@ package fr.epita.assistants.ping.presentation.rest;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import fr.epita.assistants.ping.domain.entity.UserEntity;
 import fr.epita.assistants.ping.domain.service.ProjectService;
 import fr.epita.assistants.ping.domain.service.UserService;
-import fr.epita.assistants.ping.common.api.request.UserRequest;
-import fr.epita.assistants.ping.common.api.request.AddMemberToProjectRequest;
-import fr.epita.assistants.ping.common.api.request.ExecuteFeatureRequest;
-import fr.epita.assistants.ping.common.api.request.LoginRequest;
-import fr.epita.assistants.ping.common.api.request.ProjectRequest;
-import fr.epita.assistants.ping.common.api.request.UpdateProjectRequest;
-import fr.epita.assistants.ping.common.api.response.UserResponse;
-import fr.epita.assistants.ping.common.api.request.UpdateRequest;
 import fr.epita.assistants.ping.data.model.ProjectModel;
 import fr.epita.assistants.ping.data.model.UserModel;
 import fr.epita.assistants.ping.common.api.response.GetFileResponse;
-import fr.epita.assistants.ping.common.api.response.LoginResponse;
-import fr.epita.assistants.ping.common.api.response.MemberResponse;
-import fr.epita.assistants.ping.common.api.response.ProjectResponse;
-import fr.epita.assistants.ping.common.api.response.SimpleMessageResponse;
 import fr.epita.assistants.ping.utils.ErrorInfo;
-import io.quarkus.security.Authenticated;
-import io.smallrye.jwt.build.Jwt;
-import io.smallrye.jwt.build.JwtSignature;
-import io.vertx.core.json.JsonObject;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 
+import java.nio.file.Paths;
 import java.util.*;
 
 import fr.epita.assistants.ping.common.api.request.FolderDeleteRequest;
 import fr.epita.assistants.ping.common.api.request.FolderMoveRequest;
 
 import org.eclipse.microprofile.jwt.JsonWebToken;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 
-import java.time.*;
 import fr.epita.assistants.ping.utils.Logger;
 @Path("/api/projects/{projectId}/folders")
 public class FolderResource {
@@ -51,6 +35,9 @@ public class FolderResource {
     @Inject
     JsonWebToken jwt;
 
+    @ConfigProperty(name = "PROJECT_DEFAULT_PATH", defaultValue = "")
+    String projectsPath;
+    
     @GET
     @Path("/")
     @RolesAllowed({ "admin", "user" }) // 401 + 403
@@ -88,7 +75,10 @@ public class FolderResource {
                 return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(403).build();
             }
         }
-
+        if (isPathTraversalAttack(path, id)) {
+            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "GET error attack");
+            return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE, de vouloir faire caca")).status(403).build();
+        }
         List<GetFileResponse> res = projectService.ls(id);
         Logger.logRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/",  "all ok !!!");
         return Response.ok(res).status(200).build();
@@ -107,7 +97,7 @@ public class FolderResource {
 
         ProjectModel p = projectService.getProject(id);
         if (p == null) {
-            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error p null");
+            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "DELETET error p null");
             return Response.ok(new ErrorInfo("BAKA")).status(404).build();
         }
 
@@ -132,7 +122,10 @@ public class FolderResource {
                 return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(403).build();
             }
         }
-
+        if (isPathTraversalAttack(path.relativePath, id)) {
+            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "DELETE error attack");
+            return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE, de vouloir faire caca")).status(403).build();
+        }
         if (projectService.delete(id.toString() + "/" + path.relativePath) == false) {
             Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error folder dont exist");
             return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(400).build();
@@ -179,13 +172,40 @@ public class FolderResource {
                 return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(403).build();
             }
         }
-
+        if (isPathTraversalAttack(path.relativePath, id)) {
+            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error attack");
+            return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE, de vouloir faire caca")).status(403).build();
+        }
         if (projectService.createFolder(id.toString() + "/" + path.relativePath) == false) {
             Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error folder dont exist");
             return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(409).build();
         }
         Logger.logRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/",  "all ok !!!");
         return Response.ok().status(201).build();
+    }
+
+    Boolean isPathTraversalAttack(String path, UUID id) {
+        // TODO / NOAH DONE
+        // On prend en parametre un chemin relatif
+        // par exemple:
+        // isPathTraversalAttack("/") -> C'est pas une attaque parce qu'on sort pas de
+        // root
+
+        // Il faut juste detecter si a un moment on est deja a la racine et qu'on essaye
+        // de sortir
+
+        // isPathTraversalAttack("prout/caca/../../..") -> On va sortir du projet donc
+        // on renvoie true
+
+        // Noah: Une version temporaire voir définition root selon léo
+
+        try {
+            /*TODO A modifier important*/ var root = Paths.get(projectsPath + "/" + id.toString() + "/").normalize().toAbsolutePath();
+            var target = root.resolve(path).normalize();
+            return !target.startsWith(root);
+        }catch (Exception e) {
+            return true;
+        }
     }
 
     @PUT
@@ -225,6 +245,11 @@ public class FolderResource {
                 Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error ok false");
                 return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE")).status(403).build();
             }
+        }
+
+        if (isPathTraversalAttack(path.src, id) || isPathTraversalAttack(path.dst, id)) {
+            Logger.logErrorRequest(jwt.getSubject(), "/api/projects/{projectId}/folders/", "error attack");
+            return Response.ok(new ErrorInfo("TU N'AS PAS LE DROIT ARRETE, de vouloir faire caca")).status(403).build();
         }
 
         if (projectService.moveFolder(id.toString() + "/" + path.src,id.toString() + "/" +  path.dst) == false) {
